+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Add cache summary to getCurrentState()'s Storage info
 
 ## v0.65.0
 * Add `triggerMoarData()` method to TheAlgorithm